+++ examples/overflow.py
@@ -1,0 +1,11 @@
+from typing import List
+from rich.console import Console, OverflowMethod
+
+console = Console(width=14)
+supercali = "supercalifragilisticexpialidocious"
+
+overflow_methods: List[OverflowMethod] = ["fold", "crop", "ellipsis"]
+for overflow in overflow_methods:
+    console.rule(overflow)
+    console.print(supercali, overflow=overflow, style="bold blue")
+    console.print()