+++ rich/_wrap.py
@@ -17,22 +17,26 @@ def words(text: str) -> Iterable[Tuple[int, int, str]]:
         word_match = re_word.match(text, end)
 
 
-def divide_line(text: str, width: int) -> List[int]:
+def divide_line(text: str, width: int, fold: bool = True) -> List[int]:
     divides: List[int] = []
     append = divides.append
     line_position = 0
+
     for start, _end, word in words(text):
         word_length = cell_len(word.rstrip())
         if line_position + word_length > width:
             if word_length > width:
-                for last, line in loop_last(
-                    chop_cells(word, width, position=line_position)
-                ):
-                    if last:
-                        line_position = cell_len(line)
-                    else:
-                        start += len(line)
-                        append(start)
+                if fold:
+                    for last, line in loop_last(
+                        chop_cells(word, width, position=line_position)
+                    ):
+                        if last:
+                            line_position = cell_len(line)
+                        else:
+                            start += len(line)
+                            append(start)
+                else:
+                    line_position = cell_len(word)
             elif line_position and start:
                 append(start)
                 line_position = cell_len(word)