+++ rich/syntax.py
@@ -211,8 +211,8 @@ class Syntax(JupyterMixin):
     def __rich_measure__(self, console: "Console", max_width: int) -> "Measurement":
         if self.code_width is not None:
             width = self.code_width + self._numbers_column_width
-            return Measurement(width, width)
-        return Measurement(max_width, max_width)
+            return Measurement(self._numbers_column_width, width)
+        return Measurement(self._numbers_column_width, max_width)
 
     def __rich_console__(
         self, console: Console, options: ConsoleOptions