+++ tests/test_console.py
@@ -244,3 +244,9 @@ def test_save_html():
         console.save_html(export_path)
         with open(export_path, "rt") as html_file:
             assert html_file.read() == expected
+
+
+def test_no_wrap():
+    console = Console(width=10, file=io.StringIO())
+    console.print("foo bar baz egg", no_wrap=True)
+    assert console.file.getvalue() == "foo bar ba\n"